change password, income, pay api endpoints

+++ pages/api/users/[id]/delete.js
@@ -3,24 +3,34 @@ import User from "../../../../models/User";
 import Bag from "../../../../models/Bag";
 import Order from "../../../../models/Bag";
 import { verifyToken, verifyId } from "../../../../utils/auth";
+import { compareSync } from "bcryptjs";
 
 async function handler(req, res) {
     if (req.method !== "DELETE") return res.status(405).json({ error: "DELETE method expected." });
-    const user = req.query.id;
+
+    const userId = req.query.id;
 
     try {
         await dbConnect();
-        const deletedUser = await User.findByIdAndDelete(user);
+        const user = await User.findById(userId).select("+password").lean();
 
-        if (!deletedUser) {
+        if (!user) {
             await dbDisconnect();
+
             return res
                 .status(409)
                 .json({ error: "User account not deleted because it did not exist." });
         }
 
-        await Bag.deleteMany({ user });
-        await Order.deleteMany({ user });
+        if (!compareSync(req.body.password, user.password)) {
+            await dbDisconnect();
+
+            return res.status(401).json({ error: "Incorrect password entered. Try again." });
+        }
+
+        await User.findByIdAndDelete(userId);
+        await Bag.deleteMany({ user: userId });
+        await Order.deleteMany({ user: userId });
         await dbDisconnect();
         return res.status(200).json({ message: "User account successfully deleted." });
     } catch (err) {
